Add set key

+++ libKeyWordsCompletion/Interface/ICompareKey.cs
@@ -4,6 +4,6 @@ namespace Utopia.Algo.KeyWordCompletion.Interface
 {
     public interface ICompareKey
     {
-        double Compare(IKeyWord leftWord, IKeyWord rightWord);
+        int Compare(IKeyWord leftWord, IKeyWord rightWord);
     }
 }
